chore: rename collectors to collector
This commit renames the collectors with collector for singular naming
consistency. It also updates `BuildInfoCollector` to `KeplerBuildInfoCollector`
to avoid conflicts with Prometheus `BuildInfoCollector`.
