+++ internal/exporter/prometheus/collector/power_collector_test.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2025 The Kepler Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package collectors
+package collector
 
 import (
 	"context"
